refactor(server): untangle oauth2.go into cohesive files
oauth2.go had become a grab-bag. Redistribute its contents to where they belong,
without behaviour changes:

- authorize.go gets parseAuthorizationRequest, the auth-error rendering types
  (displayedAuthErr/redirectedAuthErr), and the request validation helpers
  (validateRedirectURI/isHostLocal/validateConnectorID).
- session.go gets validateIDTokenHint, sessionMatchesHint and the signerKeySet
  verification adapter (all id_token_hint concerns).
- token.go gets tokenErr.

oauth2.go is left with the OAuth2 protocol constants, parseScopes and
validateCrossClientTrust (553 -> 114 lines).
